+++ contracts/access/events/ERC173Events.sol
@@ -6,5 +6,6 @@ pragma solidity ^0.8.22;
 /// @param newOwner the new contract owner.
 event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
 
-
-
+/// @notice Emitted when a new contract owner is pending.
+/// @param pendingOwner the address of the new contract owner.
+event OwnershipTransferPending(address indexed pendingOwner);